+++ tests/backend_api/test_connector.py
@@ -5,10 +5,13 @@ import pytest
 from requests.exceptions import HTTPError
 
 from mex.common.backend_api.connector import BackendApiConnector
-from mex.common.backend_api.models import ExtractedItemsRequest, PreviewItemsResponse
 from mex.common.models import (
+    AnyExtractedModel,
+    AnyPreviewModel,
     ExtractedPerson,
+    ItemsContainer,
     MergedPerson,
+    PaginatedItemsContainer,
     PersonRuleSetRequest,
     PersonRuleSetResponse,
     PreviewPerson,
@@ -22,16 +25,16 @@ def test_set_authentication_mocked() -> None:
     assert connector.session.headers["X-API-Key"] == "dummy_write_key"
 
 
-def test_post_extracted_items_mocked(
+def test_ingest_mocked(
     mocked_backend: MagicMock, extracted_person: ExtractedPerson
 ) -> None:
-    mocked_return = {"identifiers": [extracted_person.identifier]}
+    mocked_return = {"items": [extracted_person]}
     mocked_backend.return_value.json.return_value = mocked_return
 
     connector = BackendApiConnector.get()
-    response = connector.post_extracted_items([extracted_person])
+    response = connector.ingest([extracted_person])
 
-    assert response.identifiers == [extracted_person.identifier]
+    assert response == [extracted_person]
     assert mocked_backend.call_args == call(
         "POST",
         "http://localhost:8080/v0/ingest",
@@ -40,12 +43,12 @@ def test_post_extracted_items_mocked(
             "Accept": "application/json",
             "User-Agent": "rki/mex",
         },
-        timeout=10,
+        timeout=BackendApiConnector.INGEST_TIMEOUT,
         data=Joker(),
     )
     assert (
         json.loads(mocked_backend.call_args.kwargs["data"])
-        == ExtractedItemsRequest(items=[extracted_person]).model_dump()
+        == ItemsContainer[AnyExtractedModel](items=[extracted_person]).model_dump()
     )
 
 
@@ -202,7 +205,9 @@ def test_fetch_preview_items_mocked(
     mocked_backend: MagicMock,
     preview_person: PreviewPerson,
 ) -> None:
-    preview_response = PreviewItemsResponse(items=[preview_person], total=92)
+    preview_response = PaginatedItemsContainer[AnyPreviewModel](
+        items=[preview_person], total=92
+    )
     mocked_return = preview_response.model_dump()
     mocked_backend.return_value.json.return_value = mocked_return